# Keep a list of lazy values instead of re-traversing the whole response