In the current implementation, after eager execution is finished, the entire response is re-traversed to find any `GraphQL::Execution::Lazy` values which need to be resumed. If there _aren't_ any `Lazy`s , then this traversal is a total waste of time, because it doesn't find anything. 

So, we can improve this by using a dedicated data structure to manage resolving `Lazy` values, then traversing _that_ instead of re-using the response for this purpose. 


In my little benchmark, this shows a very good improvement in the no-Lazies case and a small improvement in the case with lots of Lazies:

### Without any lazy values 


```diff
  Calculating -------------------------------------
  Querying for 1000 objects
-                          4.379  (± 0.0%) i/s -     22.000  in   5.024509s
+                          4.628  (± 0.0%) i/s -     24.000  in   5.187606s


- Total allocated: 7155360 bytes (45479 objects)
+ Total allocated: 6995440 bytes (44480 objects)
```

### With lazy values:

```diff 
-                           3.521  (± 0.0%) i/s -     18.000  in   5.118565s
+                           3.626  (± 0.0%) i/s -     19.000  in   5.244818s

- Total allocated: 12776848 bytes (101481 objects)
+ Total allocated: 12736392 bytes (100486 objects)

```

TODO: 

- [x] Investigate failing tests. Changing the order of the response is probably ok, but annoying. 
  - In all of GraphQL-Ruby's test suite, only one test changed order. I think this is going to have to be good enough. Besides, a future GraphQL-Ruby version will properly make the response match the query shape.